Moved the logic of sort into AVector Colum #101
