+++ src/column/AVectorColumn.ts
@@ -4,7 +4,10 @@ import {IStringValueTypeDesc, IDataType} from 'phovea_core/src/datatype';
 import Range1D from 'phovea_core/src/range/Range1D';
 import {MultiForm, IMultiFormOptions} from 'phovea_core/src/multiform';
 import {list as rlist} from 'phovea_core/src/range';
+import SortColumn from './SortColumn';
+import {sort} from './SortColumn';
 import {scaleTo} from './utils';
+import * as d3 from 'd3';
 /**
  * Created by Samuel Gratzl on 19.01.2017.
  */
@@ -15,6 +18,7 @@ export abstract class AVectorColumn<T, DATATYPE extends IVector<T, any>> extends
   protected multiform: MultiForm;
   dataView: DATATYPE;
   filterRange;
+  static readonly EVENT_SORT_CHANGED = 'sorted';
 
   constructor(data: DATATYPE, orientation: EOrientation) {
     super(data, orientation);
@@ -36,6 +40,10 @@ export abstract class AVectorColumn<T, DATATYPE extends IVector<T, any>> extends
       const vislist = <HTMLElement>toolbar.querySelector('div.vislist');
       this.multiform.addIconVisChooser(vislist);
     }
+    toolbar.insertAdjacentHTML('beforeend', `<button class="fa sort fa-sort-amount-desc"></button>`);
+    const sortButton = toolbar.querySelector('button.fa-sort-amount-desc');
+
+    this.performSort(sortButton);
     super.buildToolbar(toolbar);
   }
 
@@ -57,6 +65,30 @@ export abstract class AVectorColumn<T, DATATYPE extends IVector<T, any>> extends
   //     this.multiform = this.replaceMultiForm(view, this.body);
   //   });
   // }
+
+  protected performSort(sortButton) {
+
+    sortButton.addEventListener('click', () => {
+      const b = d3.select(sortButton);
+      if (b.classed('fa-sort-amount-desc')) {
+        const sortMethod = sort.asc;
+        const s = new SortColumn((<any>this).dataView, sortMethod);
+        s.on(AVectorColumn.EVENT_SORT_CHANGED, (event: any, range) => {
+          this.fire(AVectorColumn.EVENT_SORT_CHANGED, range);
+        });
+        b.attr('class', 'fa sort fa-sort-amount-asc');
+      } else {
+        const sortMethod = sort.desc;
+        const s = new SortColumn((<any>this).dataView, sortMethod);
+        s.on(AVectorColumn.EVENT_SORT_CHANGED, (event: any, range) => {
+          this.fire(AVectorColumn.EVENT_SORT_CHANGED, range);
+
+        });
+        b.attr('class', 'fa sort fa-sort-amount-desc');
+      }
+    });
+  }
+
 
   async update(idRange: Range1D) {
     this.multiform.destroy();
