+++ src/column/ColumnManager.ts
@@ -5,7 +5,7 @@ import {
   IDataType
 } from 'phovea_core/src/datatype';
 import Range1D from 'phovea_core/src/range/Range1D';
-import {IStringVector} from './AVectorColumn';
+import {IStringVector, AVectorColumn} from './AVectorColumn';
 import AColumn, {EOrientation} from './AColumn';
 import CategoricalColumn from './CategoricalColumn';
 import StringColumn from './StringColumn';
@@ -56,7 +56,7 @@ export default class ColumnManager extends EventHandler {
     await col.update(r.intersect(this.rangeNow));
 
     col.on(AColumn.EVENT_REMOVE_ME, this.onColumnRemoved);
-    col.on(AColumn.EVENT_SORT_CHANGED, this.onColumnSort.bind(this));
+    col.on(AVectorColumn.EVENT_SORT_CHANGED, this.onColumnSort.bind(this));
     this.columns.push(col);
 
     const managerWidth = this.node.clientWidth;
@@ -90,7 +90,7 @@ export default class ColumnManager extends EventHandler {
     this.columns.splice(this.columns.indexOf(col), 1);
     col.node.remove();
     col.off(AColumn.EVENT_REMOVE_ME, this.onColumnRemoved);
-    col.off(AColumn.EVENT_SORT_CHANGED, this.onColumnSort.bind(this));
+    col.off(AVectorColumn.EVENT_SORT_CHANGED, this.onColumnSort.bind(this));
     this.fire(ColumnManager.EVENT_COLUMN_REMOVED, col);
     this.fire(ColumnManager.EVENT_DATA_REMOVED, col.data);
     this.relayout();
